Move wrapper function from parsl to base module
We are likely to use this again with FuncX (and other executors)
