+++ colmena/task_server/base.py
@@ -1,12 +1,17 @@
 """Base class for the Task Server"""
+import os
+import platform
 
 from abc import ABCMeta, abstractmethod
 from multiprocessing import Process
-from typing import Optional
+from time import perf_counter
+from typing import Optional, Callable
 import logging
 
 from colmena.exceptions import KillSignalException, TimeoutException
 from colmena.redis.queue import TaskServerQueues
+from models import Result, FailureInformation
+from proxy import resolve_proxies_async
 
 logger = logging.getLogger(__name__)
 
@@ -71,3 +76,55 @@ class BaseTaskServer(Process, metaclass=ABCMeta):
 
         # Shutdown any needed functions
         self._cleanup()
+
+
+def run_and_record_timing(func: Callable, result: Result) -> Result:
+    """Run a function and also return the runtime
+
+    Args:
+        func: Function to invoke
+        result: Result object describing task request
+    Returns:
+        Result object with the serialized result
+    """
+    # Mark that compute has started on the worker
+    result.mark_compute_started()
+
+    # Unpack the inputs
+    result.time_deserialize_inputs = result.deserialize()
+
+    # Start resolving any proxies in the input asynchronously
+    start_time = perf_counter()
+    resolve_proxies_async(result.args)
+    resolve_proxies_async(result.kwargs)
+    result.time_async_resolve_proxies = perf_counter() - start_time
+
+    # Execute the function
+    start_time = perf_counter()
+    success = True
+    try:
+        output = func(*result.args, **result.kwargs)
+    except BaseException as e:
+        output = None
+        success = False
+        result.failure_info = FailureInformation.from_exception(e)
+    finally:
+        end_time = perf_counter()
+
+    # Store the results
+    result.set_result(output, end_time - start_time)
+    if not success:
+        result.success = False
+
+    # Add the worker information into the tasks, if available
+    if result.task_info is None:
+        result.task_info = {}
+    for tag in ['PARSL_WORKER_RANK', 'PARSL_WORKER_POOL_ID']:
+        if tag in os.environ:
+            result.task_info[tag] = os.environ[tag]
+    result.task_info['executor'] = platform.node()
+
+    # Re-pack the results
+    result.time_serialize_results = result.serialize()
+
+    return result
